Reading Input Files

GenX.load_inputs — Method

load_inputs(setup::Dict,path::AbstractString)

Loads various data inputs from multiple input .csv files in path directory and stores variables in a Dict (dictionary) object for use in model() function

inputs: setup - dict object containing setup parameters path - string path to working directory

returns: Dict (dictionary) object containing all data inputs

Capacity Reserve Margin

$CO_2$ Emissions Cap

Energy Share Requirement

Fuels Data

GenX.load_fuels_data — Method

load_fuels_data(setup::Dict, path::AbstractString, sep::AbstractString, inputs_fuel::Dict)

Function for reading input parameters related to fuel costs and CO$_2$ content of fuels

Generators Input Data

GenX.load_generators_data — Method

load_generators_data(setup::Dict, path::AbstractString, sep::AbstractString, inputs_gen::Dict, fuel_costs::Dict, fuel_CO2::Dict)

Function for reading input parameters related to electricity generators (plus storage and flexible demand resources)

Variability of Generators' Outputs

GenX.load_generators_variability — Method

load_generators_variability(setup::Dict, path::AbstractString, sep::AbstractString, inputs_genvar::Dict)

Function for reading input parameters related to hourly maximum capacity factors for all generators (plus storage and flexible demand resources)

Load Data

GenX.load_load_data — Method

load_data(setup::Dict, path::AbstractString, sep::AbstractString, inputs_load::Dict)

Function for reading input parameters related to electricity load (demand)

Minimum Capacity Requirements

GenX.load_minimum_capacity_requirement — Method

load_minimum_capacity_requirement(path::AbstractString,sep::AbstractString, inputs::Dict, setup::Dict)

Function for reading input parameters related to mimimum capacity requirement constraints (e.g. technology specific deployment mandates)

Transmission Network

GenX.load_network_data — Method

load_network_data(setup::Dict, path::AbstractString, sep::AbstractString, inputs_nw::Dict)

Function for reading input parameters related to the electricity transmission network

Mapping Representative Time Periods

GenX.load_period_map — Method

load_period_map(setup::Dict,path::AbstractString,sep::AbstractString, inputs::Dict)

Function for reading input parameters related to mapping of representative time periods to full chronological time series

Reserves Data

GenX.load_reserves — Method

load_reserves(setup::Dict,path::AbstractString,sep::AbstractString, inputs_res::Dict)

Function for reading input parameters related to frequency regulation and operating reserve requirements